core,metautils: Factorize the DJB hash functions
