+++ metautils/lib/metautils.h
@@ -34,7 +34,6 @@ License along with this library.
 
 # include <metautils/lib/metatypes.h>
 
-# include <metautils/lib/metautils_bits.h>
 # include <metautils/lib/metautils_errors.h>
 # include <metautils/lib/metautils_strings.h>
 # include <metautils/lib/metautils_sockets.h>
